+++ MercadoPagoSDK/MercadoPagoSDK/One tap/UI/PXOneTapPaymentMethodView.swift
@@ -1,0 +1,15 @@
+//
+//  PXOneTapPaymentMethodView.swift
+//  MercadoPagoSDK
+//
+//  Created by Juan sebastian Sanzone on 16/5/18.
+//  Copyright © 2018 MercadoPago. All rights reserved.
+//
+
+import Foundation
+
+final class PXOneTapPaymentMethodView: PXBodyView {
+    var paymentMethodIcon: UIView?
+    var paymentMethodTitle: UILabel?
+    var paymentMethodSubtitle: UILabel?
+}
