+++ MercadoPagoSDK/MercadoPagoSDK/PXFooterView.swift
@@ -1,0 +1,14 @@
+//
+//  PXFooterView.swift
+//  MercadoPagoSDK
+//
+//  Created by Juan sebastian Sanzone on 16/5/18.
+//  Copyright © 2018 MercadoPago. All rights reserved.
+//
+
+import Foundation
+
+class PXFooterView: UIView {
+    public var principalButton: PXPrimaryButton?
+    public var linkButton: PXSecondaryButton?
+}
